Move generated JIT autograd tests to test_jit
